+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix account chart issues #421
 - Fix modal title and data card spacing in accounts #426
 - Fix hash on account page to draw from utility_token ednpoint #431
+- Fix transactions history display when data length is 0 #436
 
 ### Features
 - Add data download csv for transaction volume #399